Refactored test into separate class, to avoid needless repeated execution by subclasses of the original test.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/lang/trunk@1673049 13f79535-47bb-0310-9956-ffa450edef68
