+++ src/test/java/org/apache/commons/lang3/time/FastDateParserTest.java
@@ -21,7 +21,6 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import java.io.Serializable;
-import java.text.DateFormatSymbols;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
@@ -31,7 +30,6 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 import java.util.TimeZone;
-import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.SerializationUtils;
 import org.junit.Assert;
@@ -641,20 +639,4 @@ public class FastDateParserTest {
             assertEquals(message+trial.three, cal.getTime(), parser.parse(dateStub+trial.three));
         }
     }
-
-    @Test
-    public void testTimeZoneStrategyPattern() {
-        Pattern tz = Pattern.compile(FastDateParser.TimeZoneStrategy.TZ_DATABASE);
-        Assert.assertFalse(tz.matcher("GMT-1234").matches());
-
-        for (Locale locale : Locale.getAvailableLocales()) {
-            final String[][] zones = DateFormatSymbols.getInstance(locale).getZoneStrings();
-            for (final String[] zone : zones) {
-                for (String zoneExpr : zone) {
-                    Assert.assertTrue(locale.getDisplayName() + ":" + zoneExpr, tz.matcher(zoneExpr).matches());
-                }
-            }
-        }
-    }
-
 }
